define file structure, build server-client connection (separate server and client)
