+++ src/main/java/edu/sjsu/cmpe275/team6/SnippetShare/main/JpaTest.java
@@ -1,0 +1,7 @@
+package edu.sjsu.cmpe275.team6.SnippetShare.main;
+
+/**
+ * Created by Corn on 4/6/15.
+ */
+public class JpaTest {
+}
