+++ src/main/java/edu/sjsu/cmpe275/team6/SnippetShare/model/User.java
@@ -1,0 +1,7 @@
+package edu.sjsu.cmpe275.team6.SnippetShare.model;
+
+/**
+ * Created by Corn on 4/6/15.
+ */
+public class User {
+}
