+++ src/main/webapp/WEB-INF/pages/index.jsp
@@ -1,0 +1,16 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: Corn
+  Date: 4/6/15
+  Time: 3:32 PM
+  To change this template use File | Settings | File Templates.
+--%>
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<html>
+<head>
+    <title></title>
+</head>
+<body>
+
+</body>
+</html>
