+++ src/main/webapp/css/style.css
@@ -1,0 +1,3 @@
+body {
+    background-color: lightgrey;
+}
