+++ src/main/webapp/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html>
+<head lang="en">
+    <meta charset="UTF-8">
+    <title></title>
+    <link rel="stylesheet" href="css/style.css"/>
+</head>
+<body>
+<h2>Home page test</h2>
+
+
+
+</body>
+</html>
