+++ src/test/java/edu/sjsu/cmpe275/team6/SnippetShare/AppTests.java
@@ -17,7 +17,7 @@ import static org.springframework.test.web.servlet.setup.MockMvcBuilders.webAppC
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @WebAppConfiguration
-@ContextConfiguration("file:src/main/webapp/WEB-INF/mvc-dispatcher-servlet.xml")
+@ContextConfiguration("file:../../../../../../../main/webapp/WEB-INF/snippetshare-servlet.xml")
 public class AppTests {
     private MockMvc mockMvc;
